Reoranize clinic invite implementation
