+++ go.mod
@@ -9,7 +9,7 @@ require (
 	github.com/gorilla/mux v1.8.0
 	github.com/kelseyhightower/envconfig v1.4.0
 	github.com/pkg/errors v0.9.1
-	github.com/tidepool-org/clinic/client v0.0.0-20210322141342-076ebb0f702c
+	github.com/tidepool-org/clinic/client v0.0.0-20210330122713-69bf15471908
 	github.com/tidepool-org/go-common v0.7.1
 	go.mongodb.org/mongo-driver v1.4.1
 	go.uber.org/fx v1.13.1
